xobixiangqing/backend/requirements.txt
# Runtime dependencies mirror pyproject.toml
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.1
google-genai>=1.52.0
openai>=1.0.0
pydantic>=2.9.0
pillow>=12.0.0
python-dotenv>=1.0.1
werkzeug>=3.0.1
tenacity>=9.0.0
alembic>=1.13.0
flask-migrate>=4.0.0
openpyxl>=3.1.2
xlrd>=2.0.1
httpx>=0.27.0
